Fix bug with client and lack of package name
Within our client, when creating a new GRPC-transport-client, we where
specifying a Service name that didn't include a package name, only the
bare name of the Service. go-kit has some odd behavior in a case where
a service name doesn't include a '.' character, which is the case when
the package name isn't included. The odd behavior is that it will
prepend the service name you give it with the string "pb.", basically
assuming that the package name that the service lives in is "pb".

Sadly, this assumption isn't documented in go-kit anywhere, so it took
us some time to find the source, but this commit fixes the issue!