Add unit test for makedt
The addition of annotations.pb.go and http.pb.go are necessary so that
tests which make use of http annotations are able to make use of them.
They're not used in this single test, but for future tests it's
important that they be included so I'm including them now.