Major truss refactor
truss used to be composed of several protoc plugins, with lots of side
effects.

truss is now composed of one protoc plugin, protoc-gen-truss-protocout
which outputs the output of protoc.

This output, plus the .proto file with the service definition are used
to create a doctree

This doctree is then used by gendocs and gengokit to generate the
microservice and documentation.

This stucture has less side effects, and should be more testable