Update integration tests with http transport
Before the integration tests were only testing building the server and
cli client and testing for panics over grpc.

Now the following is tested:
- http transport, both the client library and manually set http requests
    - query, repeated query, and nested messages in body
- building cliclient and server
    - gprc and http transports
    - base, repeated, nested messages, and map type tests

The repeated, nested messages, and map type tests are currently disabled
until those features are implmented for the cliclient.

In addition a Makefile to the root directory has been added for building
truss and running all of the tests.