Add support for govendor in addtion for glide.
Proposal:
Switch over to glide entirely. I switched my fork to be managed with
govendor which I find friendlier. I think it would be nice to just
switch truss over as well. If we wanna go that route then we can merge
this and follow up with a second PR to remove the glide files so we
don't get out of sync.