Update everything to use .pb.go interface
Before in NAME-service/handlers/server there was a Service interface
that everything used. This caused issues because if you imported a
client, it would import the handlers, which would import basically the
whole service.

By changing to the .pb.go interface, this is no longer an issue.
Additionally there is code simplification in gengokit.go as the client
and server handler method updating functions have now merged and
partial_templates have been removed. The flow in gengokit is currently
not the greatest, though it is being refactored shortly.