Remove the truss.NamedReadWriter interface and truss.SimpleFile struct.
No functional change

This interface was originally because we hoped that os.File would satisfy this
interface, though the Name() method of os.File would not work correctly.

It is replaced with a map[string]io.Reader which represents a relative path on
disk to the service directory to file contents. It is used as the structure for
both reading in files and how code generating packages output files to write.