Add SHA and build date to Truss binary when built with Makefile
With Truss as an ever-moving target, it may be useful to know what
version you are running. This change will embed the Git SHA and the
build date into the Truss binary, if Truss is built with the Makefile.
If the SHA or build date are set in the binary, it will be displayed in
the usage message. If no SHA or build date are set, no banner will be
displayed.

Note that the Makefile is not used when fetching with `go get` or
building a fetched source repo with `go install`, so binaries built like
this will not have the SHA and build metadata embedded.