Add additional endpoint middleware wrapping function
As it was, InjectEndpointMiddlewares was called for every endpoint on
the defined service. This is fine so long as you want these middlewares
applied to every single endpoint. This breaks down, though, when your
middleware desires are heterogeneous, like, for example, protecting
some endpoints from free access by enforcing some sort of authentication
and authorization, while allowing other endpoints to be freely
accessible (e.g., login, service status, ping).

This change introduces another function into the endpoint middlewares
template, which takes the entire set of endpoints as defined for the
service and returns another set of endpoints. This way, the programmer
has the flexibility wrap different endpoints differently.

This is how service middlewares work, since the middleware wrapping
function takes a service interface, and returns a service interface, but
we should not force a programmer to use Service middlewares if a
middleware would more appropriately be an endpoint middleware.

Additionally, the naming of the middleware application functions has
been changed from using the verb "inject" to the verb "wrap," which
makes more sense in the context of middlewares, which wrap a deeper
functionality.