[#131] Supress warnings when httpoptions are not defined
Before if the user did not specify any httpoptions truss would spew a
rather lengthly warning about how http transport would not work. This
change removes those warnings under the assumption if httpoptions are
missing the users probably did so intentionally.

In the case where the annotations needed to parse the httpoptions is not
imported and httpoptions exist. Truss follows existing behavior and dies
immediately.