Use PBFieldName (json tag name) as fieldname in query string
Closes #12

When using the exposed HTTP API of a Truss generated web service the
query parameter names were required to be in PascalCase. This made sense
thinking about the underlying Go data structure, but not when working
with HTTP / QueryParams. Likewise its more consistent if
instead the snake cased names are used as they will match the json
fields that are accepted.