(decode) Remove unneed dictionary creation, only check error when present
from
```
	InEchoStr := queryParams["In"]
	InEcho := InEchoStr
	// TODO: Better error handling
	if err != nil {
		fmt.Printf("Error while extracting InEcho from query: %v\n", err)
		fmt.Printf("queryParams: %v\n", queryParams)
		return nil, err
	}
	req.In = InEcho

```
to
```
	queryParams := r.URL.Query()
	_ = queryParams

	if InEchoStrArr, ok := queryParams["In"]; ok {
		InEchoStr := InEchoStrArr[0]
		InEcho := InEchoStr
		req.In = InEcho
	}
```

As an added bonus, queryParams will no longer override body parameters
when not present

Additionally the error logic in the top block was always generated even
if the InEchoStr was of type string and thus passed through directly
(did not need to be parsed with ex. ParseBool). The code was unreachable
and will now nolonger be generated