Fix bug reported in issue #202
The bug is that if you attempt to associate a comment with something
that deftree doesn't have an internal representation for, it would fail
to find anything after searching for where to put the comment, and would
panic.

The fix was to exchange the panic with returning an error which can be
handled/ignored.