Remove some exposed functions from svc/server/run.go
After talking with @hasLeland it was pointed out that exposing more api
surface in the generated code is probably not the best plan. I had been
doing it to make using a custom `main.go` easier.

I believe that `server.NewEndpoints()` was a good addition to the api
because there is generated code in that function that would be a pain to
maintain from a custom `main.go`.

Though creating the `RunHTTP` and etc functions was a mistake, should
have proposed it first. They expose more api and force us into
supporting without adding comparable appropriate benefits.