fix regressions in how json is returned from the http path
after bringing in some recent changes to use jsonpb, we introduced an accidental regression where the protobuf names stopped being returned as defined in the proto. Truss had changed in place to set the json tags to the correct value, but jsonpb doesnt seem to use those and goes by the protobuf mappings themselves. Additionally we accidentally broken the omitempty behavior, by mistaking the polarity of the EmitDefaults value (oops)