Move ClientOptions to go-kit ClientOptions
By removing our old truss-specific ClientOptions and instead using
standard go-kit ClientOptions, this allows a developer to leverage more
out-of-the-box functionality without needing to adapt truss to support
those options.

This change was made to allow a custom http.Client to be provided to
New(), allowing for a different client that http.DefaultClient.